[REF] pulso: Keep the same reference when reinvoicing
Before this change, when an invoice was replaced, a new invoice code need
to be provided for the new invoice. This is not as convenient for bank
reconciliations, because the same order code is used (our invoice
reference).

After this change, all documents have the same reference: old invoice,
credit note and new invoice.