[IMP] Pulso: MX Sign flow for group invoices
The invoices created by agents are not going to send sign, but they will
be in a grouped way or they will be signed normally depending of:

Every B2C customer (Non-Agent) that generates an invoice must be sent sign and will not be grouped.
Any invoice (reservation) from an Agent will not be sent sign.
A) The only invoice that will be sent to sign is the grouping invoice.
B) Exception: When an invoice that is not the grouped one and its payment method
is Credit Card ("l10n_mx_edi_payment_method_id": "04") it must be sent sign and not grouped.
Task#56792